refactor: ResponsePage 패키지 변경
